refactor(go): remove rspress adapter shims from example app
The go-web-decouple refactor made the source-level adapter pattern
available via GoConfig. The example app now passes useI18n, useLang,
useDark, and CodeBlock directly through GoConfig instead of relying
on rsbuild module aliases and StandaloneRuntimeProvider.
